Introduced vex::Context class
These commonly met lines:
    std::vector<cl::Context>      context;
    std::vector<cl::CommandQueue> queue;
    std::tie(context, queue) = queue_list(SomeFilter);
may now be replaced with
    vex::Context ctx(SomeFilter);

It may be output to a stream, and it has three methods: context(), queue(), and
size(). First two of them return list of cl::Contexts and cl::CommandQueues,
and size() returns size of queue list.

Also note that All, Env, and DoublePecision filters are now represented by
static objects, so that you have to write
    vex::Context ctx(Filter::All)
instead of
    vex::Context ctx(Filter::All())