Simplify definition of new terminals
Kernel generation driver functions now differentiate between terminals
that are values and terminals that have values (vector<T> is both a
terminal and a value, but proto::terminal<int> is a terminal that holds
a value).

Helper functions for kernel generation are now called with appropriate
arguments (either terminal itself, or proto::value(terminal).