Pick fixed name for tagged terminals
Tagged terminals may be found in several parts of the expression. The
name for the tagged terminal used to depend on the first time it was
encountered. See #67 for an example of where it may lead to problems.
This commit addresses the issue by picking a fixed prefix for the
parameter names of the tagged terminal.