Use CUSPARSE matrices with CUDA backend
If CUDA backend is in use and VEXCL_USE_CUSPARSE macro is defined, then
vex::SpMat<> class is based on CUSPARSE library. There are both positive
and negative sides: on the plus side, matrix construction (and  probably
multiplication) happens faster; the disadvantage is that possibility to
inline matrix-vector product is lost.

Closes #78