Implemented tensor dot product operation
Given two tensors (arrays of dimension greater than or equal to one),
A and B, and a list of axes pairs (where each pair represents corresponding
axes from two tensors), sums the products of A's and B's elements over the
given axes.

Inspired by python's numpy.tensordot operation.

For example, with tensordot a dense matrix-matrix product may be represented as:

    slicer<2> Adim(extents[N][M]);
    slicer<2> Bdim(extents[M][L]);

    // C = A * B
    C = tensordot(Adim[_](A), Bdim[_](B), axes_pairs(1, 0));

This should perform better than the following combination of reshape and
reduce (see #31):

    C = reduce<SUM>(
                extents[N][M][L],
                reshape(A, extents[N][M][L], extents[0][1]) *
                reshape(B, extents[N][M][L], extents[1][2]),
                extents[1]
                );