Adding Boost.Compute backend
Core functionality of the Boost.Compute library is used as a replacement
to Khronos C++ API which seems to become more and more outdated. The
Boost.Compute backend is still based on OpenCL, so there are two OpenCL
backends now.

Define VEXCL_BACKEND_COMPUTE to use this backend and make sure
Boost.Compute headers are in include path.